backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-bcrypt==0.7.1
flask-jwt-extended==4.3.1
mysql-connector-python==8.0.26
python-dotenv==0.19.1
Werkzeug==2.0.3